full typing of mapchete.io.vector module
